them chuc nang xoa ctdt

+++ src/main/java/com/example/qlmc/repository/CTDTRepository.java
@@ -11,7 +11,7 @@ import com.example.qlmc.entity.CTDT;
 
 import jakarta.transaction.Transactional;
 
-public interface CTDTRepository extends JpaRepository<CTDT, Long> {
+public interface CTDTRepository extends JpaRepository<CTDT, Integer> {
     @Query(value = "SELECT * FROM chuongtrinhdaotao WHERE ma_kdcl = :maKdcl", nativeQuery =  true)
     List<CTDT> findAllByMaKDCL(@Param("maKdcl") String maKdcl);
 
@@ -43,6 +43,8 @@ public interface CTDTRepository extends JpaRepository<CTDT, Long> {
     @Transactional
     @Query (value = "UPDATE chuongtrinhdaotao SET ten_ctdt = :tenCtdt, ma_khoa = :maKhoa, ma_nganh = :maNganh WHERE id_ctdt = :idCtdt", nativeQuery = true)
     void updateCTDT(@Param("tenCtdt") String tenCtdt, @Param("maKhoa") String maKhoa, @Param("maNganh") String maNganh, @Param("idCtdt") int idCtdt);
+    
+
